+++ tests/KingmakerGunslinger.DomainTests/Program.cs
@@ -25,6 +25,9 @@ namespace KingmakerGunslinger.DomainTests
     {
         private static readonly TestCase[] Cases =
         {
+            Case("teleportation.layout.NativeActionsAndRowsFitAroundSelectedAnchor", TeleportationLayoutTests.NativeActionsAndRowsFitAroundSelectedAnchor),
+            Case("teleportation.layout.LongListsScrollBeforeObscuringNativeActions", TeleportationLayoutTests.LongListsScrollBeforeObscuringNativeActions),
+            Case("teleportation.layout.UnprovenOrInsufficientGeometryFailsClosed", TeleportationLayoutTests.UnprovenOrInsufficientGeometryFailsClosed),
             Case("teleportation.presentation.PreparedRowIncludesCurrentCount", TeleportationContextTests.PreparedRowIncludesCurrentCount),
             Case("teleportation.presentation.SpontaneousRowsIncludeCorrectLevelAndPlural", TeleportationContextTests.SpontaneousRowsIncludeCorrectLevelAndPlural),
             Case("teleportation.presentation.AmbiguousCasterRowsNameTheBook", TeleportationContextTests.AmbiguousCasterRowsNameTheBook),